backend/requirements.txt
fastapi
uvicorn
sqlmodel
asyncpg
python-dotenv
requests
google-api-python-client
google-auth
cryptography